refactor(reflections): split bundles into one-file-per-reflection under reflections/{group}/ (#1028)
- 19 self-contained per-reflection modules under reflections/{agents,housekeeping,audits,memory}/
- old bundles (maintenance/auditing/task_management/memory_management) become documented re-export shims
- agent/sustainability.py 5 reflections relocated to reflections/agents/; shim keeps send_hibernation_notification + _get_project_key/_get_redis defined in place
- reflections/utils.py renamed to reflections/utilities.py (git mv); all importers updated
- zero vault YAML edit needed: all historical callable dotted paths resolve via shims
- verified: all 30 enabled registry callables resolve via scheduler _resolve_callable
