+++ agent/sustainability.py
@@ -1,37 +1,49 @@
-"""Sustainable self-healing reflections for queue governance.
+"""Compatibility surface for self-healing reflections relocated to reflections/agents/.
 
-Seven standalone reflection functions, each registered in config/reflections.yaml:
-- circuit_health_gate: pause/resume queue and manage hibernation based on Anthropic circuit state
-- session_recovery_drip: drip paused_circuit and paused sessions back to pending one at a time
-- session_count_throttle: throttle sessions per hour to prevent runaway execution
-- failure_loop_detector: deduplicate GitHub issues for repeated error patterns
-- sustainability_digest: daily health check, exception-only Telegram delivery (silent when nominal)
-- send_hibernation_notification: enqueue a Telegram notification for hibernation entry or wake
+The five self-healing reflection callables that used to live here now each have
+their own self-contained module under ``reflections/agents/`` (one file per
+reflection — see issue #1028). The reflections registry (config/reflections.yaml,
+vault) still references the historical dotted paths below, so each reflection is
+re-exported here under its original name and the scheduler's importlib resolution
+keeps working with no registry edit:
 
-All functions are synchronous (run in executor by the reflection scheduler).
-All functions catch all exceptions and log — never crash the reflection tick.
+- ``agent.sustainability.circuit_health_gate``    → reflections.agents.circuit_health_gate.run
+- ``agent.sustainability.session_recovery_drip``  → reflections.agents.session_recovery_drip.run
+- ``agent.sustainability.session_count_throttle`` → reflections.agents.session_count_throttle.run
+- ``agent.sustainability.failure_loop_detector``  → reflections.agents.failure_loop_detector.run
+- ``agent.sustainability.sustainability_digest``  → reflections.agents.system_health_digest.run
 
-Redis key schema:
-- {project_key}:sustainability:queue_paused — set when circuit is OPEN/HALF_OPEN (TTL 3600s)
-- {project_key}:worker:hibernating — set when worker enters hibernation (TTL 600s)
-- {project_key}:recovery:active — set when circuit recovers from queue_paused state (TTL 3600s)
-- {project_key}:worker:recovering — set when circuit recovers from hibernation (TTL 3600s)
+New code should import the reflection directly from its per-reflection module.
 
-The {project_key} prefix resolves to ``valor`` in production, sourced from the
-``VALOR_PROJECT_KEY`` env var injected into worker/bridge launchd plists by
-the install scripts (``scripts/install_worker.sh`` and
-``scripts/update/service.py::install_worker``). When unset/empty, the fallback
-is ``"valor"`` (matching ``tools.agent_session_scheduler.DEFAULT_PROJECT_KEY``)
-to keep readers and AgentSession writers aligned. See issue #1171.
+``send_hibernation_notification`` is NOT a reflection — it is a helper imported
+directly by ``agent/agent_session_queue.py`` (the circuit-health hibernation
+path). It stays defined here, along with the ``_get_project_key`` / ``_get_redis``
+helpers that several tests and the relocated reflections rely on as the canonical
+project-key resolver.
 """
 
-import hashlib
 import logging
 import os
-import subprocess
-import time
+
+# Re-exports so config/reflections.yaml's historical callable paths still resolve.
+from reflections.agents.circuit_health_gate import run as circuit_health_gate
+from reflections.agents.failure_loop_detector import run as failure_loop_detector
+from reflections.agents.session_count_throttle import run as session_count_throttle
+from reflections.agents.session_recovery_drip import run as session_recovery_drip
+from reflections.agents.system_health_digest import run as sustainability_digest
 
 logger = logging.getLogger(__name__)
+
+__all__ = [
+    "circuit_health_gate",
+    "session_recovery_drip",
+    "session_count_throttle",
+    "failure_loop_detector",
+    "sustainability_digest",
+    "send_hibernation_notification",
+    "_get_project_key",
+    "_get_redis",
+]
 
 
 def _get_project_key() -> str:
@@ -51,164 +63,6 @@ def _get_redis():
     from popoto.redis_db import POPOTO_REDIS_DB
 
     return POPOTO_REDIS_DB
-
-
-def circuit_health_gate() -> None:
-    """Check Anthropic circuit state and manage all circuit-related flags atomically.
-
-    Replaces api_health_gate (sustainability.py) and worker_health_gate (hibernation.py).
-
-    - OPEN or HALF_OPEN → renew queue_paused (TTL 3600s) AND worker:hibernating (TTL 600s)
-    - CLOSED → delete both flags; if either was set, set recovery:active AND worker:recovering
-      (TTL 3600s each) and call send_hibernation_notification("waking")
-
-    Logs at WARNING on first transition to hibernated/paused state.
-    Guards against cold-start / test environments where circuit is not registered.
-    Never raises — all exceptions are caught and logged.
-    """
-    try:
-        from bridge.health import get_health
-        from bridge.resilience import CircuitState
-
-        r = _get_redis()
-        project_key = _get_project_key()
-        pause_key = f"{project_key}:sustainability:queue_paused"
-        hib_key = f"{project_key}:worker:hibernating"
-        recovery_key = f"{project_key}:recovery:active"
-        rec_key = f"{project_key}:worker:recovering"
-
-        cb = get_health().get("anthropic")
-        if cb is None:
-            logger.debug("[circuit-health-gate] Anthropic circuit not registered — skipping")
-            return
-
-        if cb.state == CircuitState.CLOSED:
-            # Circuit closed — clear both flags atomically
-            was_paused = r.exists(pause_key)
-            was_hibernating = r.exists(hib_key)
-            r.delete(pause_key)
-            r.delete(hib_key)
-            was_either_flag_set = was_paused or was_hibernating
-            if was_either_flag_set:
-                logger.info(
-                    "[circuit-health-gate] Anthropic circuit CLOSED — queue unpaused,"
-                    " hibernation cleared, starting recovery drip"
-                )
-                r.set(recovery_key, "1", ex=3600)
-                r.set(rec_key, "1", ex=3600)
-                # Enqueue wake notification (best-effort)
-                try:
-                    send_hibernation_notification("waking", project_key=project_key)
-                except Exception as _notif_err:
-                    logger.error(
-                        "[circuit-health-gate] Failed to enqueue wake notification: %s", _notif_err
-                    )
-            else:
-                logger.debug(
-                    "[circuit-health-gate] Anthropic circuit CLOSED — queue was not paused,"
-                    " worker was not hibernating"
-                )
-        else:
-            # OPEN or HALF_OPEN — renew both flags
-            was_paused = r.exists(pause_key)
-            was_hibernating = r.exists(hib_key)
-            r.set(pause_key, "1", ex=3600)
-            r.set(hib_key, "1", ex=600)
-            if not was_paused and not was_hibernating:
-                logger.warning(
-                    "[circuit-health-gate] Queue paused, worker hibernating — Anthropic circuit %s",
-                    cb.state.value.upper(),
-                )
-            else:
-                logger.debug(
-                    "[circuit-health-gate] Queue remains paused, worker remains hibernating"
-                    " — Anthropic circuit %s",
-                    cb.state.value.upper(),
-                )
-    except Exception:
-        logger.exception("[circuit-health-gate] Unhandled exception — skipping tick")
-
-
-def session_recovery_drip() -> None:
-    """Drip one paused session back to pending per tick.
-
-    Replaces recovery_drip (sustainability.py) and session_resume_drip (hibernation.py).
-
-    Only active when recovery:active OR worker:recovering flag is set in Redis.
-    Drips paused_circuit sessions first (FIFO), then paused sessions (FIFO).
-    Clears BOTH recovery:active AND worker:recovering when both queues are empty.
-
-    Rate: called every 30s by the scheduler → at most 1 session per 30s.
-    Never raises — all exceptions are caught and logged.
-    """
-    try:
-        from agent.session_health import _filter_hydrated_sessions
-        from models.agent_session import AgentSession
-        from models.session_lifecycle import transition_status
-
-        r = _get_redis()
-        project_key = _get_project_key()
-        recovery_key = f"{project_key}:recovery:active"
-        rec_key = f"{project_key}:worker:recovering"
-
-        if not r.exists(recovery_key) and not r.exists(rec_key):
-            logger.debug("[session-recovery-drip] neither recovery flag set — no-op")
-            return
-
-        # Phantom guard: drop records whose fields are still Popoto Field descriptors
-        # (orphan $IndexF members).
-        paused_circuit = _filter_hydrated_sessions(
-            AgentSession.query.filter(project_key=project_key, status="paused_circuit")
-        )
-        paused = _filter_hydrated_sessions(
-            AgentSession.query.filter(project_key=project_key, status="paused")
-        )
-
-        if not paused_circuit and not paused:
-            r.delete(recovery_key)
-            r.delete(rec_key)
-            logger.info(
-                "[session-recovery-drip] both queues empty — clearing recovery:active"
-                " and worker:recovering flags"
-            )
-            return
-
-        # Pop oldest session (FIFO by created_at) — paused_circuit has priority
-        def _ts(session):
-            from bridge.utc import to_unix_ts
-
-            return to_unix_ts(getattr(session, "created_at", None)) or 0.0
-
-        if paused_circuit:
-            paused_circuit.sort(key=_ts)
-            candidate = paused_circuit[0]
-            drip_reason = "session-recovery-drip: API circuit recovered"
-        else:
-            paused.sort(key=_ts)
-            candidate = paused[0]
-            drip_reason = "session-recovery-drip: worker recovered"
-
-        remaining = len(paused_circuit) + len(paused) - 1
-
-        try:
-            transition_status(
-                candidate,
-                "pending",
-                reason=drip_reason,
-            )
-            logger.info(
-                "[session-recovery-drip] Dripped session %s → pending (%d remaining paused)",
-                getattr(candidate, "session_id", "?"),
-                remaining,
-            )
-        except Exception as e:
-            logger.warning(
-                "[session-recovery-drip] Could not transition session %s to pending: %s",
-                getattr(candidate, "session_id", "?"),
-                e,
-            )
-    except Exception:
-        logger.exception("[session-recovery-drip] Unhandled exception — skipping tick")
 
 
 def send_hibernation_notification(event: str, project_key: str | None = None) -> None:
@@ -270,405 +124,3 @@ def send_hibernation_notification(event: str, project_key: str | None = None) ->
         )
     except Exception as e:
         logger.error("[circuit-health-gate] Failed to enqueue %s notification: %s", event, e)
-
-
-def session_count_throttle() -> None:
-    """Count sessions started in the last hour and write throttle level.
-
-    Throttle levels:
-    - none: sessions/hr < SUSTAINABILITY_THROTTLE_MODERATE
-    - moderate: sessions/hr >= SUSTAINABILITY_THROTTLE_MODERATE (blocks low-priority only)
-    - suspended: sessions/hr >= SUSTAINABILITY_THROTTLE_SUSPENDED (blocks normal + low)
-
-    Thresholds are configurable via env vars (defaults: 20 and 40).
-    """
-    try:
-        from agent.session_health import _filter_hydrated_sessions
-        from models.agent_session import AgentSession
-
-        r = _get_redis()
-        project_key = _get_project_key()
-        throttle_key = f"{project_key}:sustainability:throttle_level"
-
-        moderate_threshold = int(os.environ.get("SUSTAINABILITY_THROTTLE_MODERATE", "20"))
-        suspended_threshold = int(os.environ.get("SUSTAINABILITY_THROTTLE_SUSPENDED", "40"))
-
-        cutoff = time.time() - 3600  # 1 hour ago
-        # Phantom guard: drop records whose fields are still Popoto Field descriptors
-        # (orphan $IndexF members).
-        all_sessions = _filter_hydrated_sessions(AgentSession.query.filter(project_key=project_key))
-        from bridge.utc import to_unix_ts
-
-        recent = []
-        for s in all_sessions:
-            ts = to_unix_ts(getattr(s, "started_at", None))
-            if ts is None:
-                continue
-            if ts >= cutoff:
-                recent.append(s)
-
-        count = len(recent)
-        if count >= suspended_threshold:
-            level = "suspended"
-        elif count >= moderate_threshold:
-            level = "moderate"
-        else:
-            level = "none"
-
-        r.set(throttle_key, level, ex=7200)  # TTL 2hr (survive across two hourly ticks)
-        logger.info(
-            "[session-count-throttle] %d sessions started in last hour → throttle_level=%s",
-            count,
-            level,
-        )
-    except Exception:
-        logger.exception("[session-count-throttle] Unhandled exception — skipping tick")
-
-
-def failure_loop_detector() -> None:
-    """Detect repeated error fingerprints and file one GitHub issue per novel cluster.
-
-    Scans failed sessions in the last 4 hours. Groups by error fingerprint.
-    For each cluster with >= 3 failures:
-    - Checks Redis set for seen fingerprints (SADD, skip if return value 0)
-    - Files one GitHub issue via gh CLI
-    - Marks affected sessions with loop_detected=True in extra_context
-
-    Skips analysis entirely if the queue is paused (API outage in progress).
-    """
-    try:
-        from agent.session_health import _filter_hydrated_sessions
-        from models.agent_session import AgentSession
-
-        r = _get_redis()
-        project_key = _get_project_key()
-        pause_key = f"{project_key}:sustainability:queue_paused"
-        seen_key = f"{project_key}:sustainability:seen_fingerprints"
-
-        # Skip during acknowledged API outages
-        if r.exists(pause_key):
-            logger.info("[failure-loop-detector] Queue paused (API outage) — skipping failure scan")
-            return
-
-        from bridge.utc import to_unix_ts
-
-        cutoff = time.time() - (4 * 3600)  # 4 hours ago
-        # Phantom guard: drop records whose fields are still Popoto Field descriptors
-        # (orphan $IndexF members).
-        all_sessions = _filter_hydrated_sessions(AgentSession.query.filter(project_key=project_key))
-        failed_sessions = []
-        for s in all_sessions:
-            if s.status not in ("failed", "abandoned"):
-                continue
-            ts = to_unix_ts(getattr(s, "completed_at", None))
-            if ts is None:
-                continue
-            if ts >= cutoff:
-                failed_sessions.append(s)
-
-        if not failed_sessions:
-            logger.debug("[failure-loop-detector] No failed sessions in last 4 hours")
-            return
-
-        # Build fingerprint clusters
-        clusters: dict[str, list] = {}
-        for s in failed_sessions:
-            fingerprint = _compute_fingerprint(s)
-            clusters.setdefault(fingerprint, []).append(s)
-
-        for fingerprint, sessions in clusters.items():
-            if len(sessions) < 3:
-                continue
-
-            # Atomic check-and-set: SADD returns 1 if new, 0 if already existed
-            added = r.sadd(seen_key, fingerprint)
-            if r.ttl(seen_key) < 0:
-                # Set TTL if not already set (7 days)
-                r.expire(seen_key, 7 * 86400)
-
-            if added == 0:
-                logger.debug(
-                    "[failure-loop-detector] Fingerprint %s already seen — skipping issue creation",
-                    fingerprint,
-                )
-                continue
-
-            # File GitHub issue
-            session_ids = [getattr(s, "session_id", "?") for s in sessions[:5]]
-            _file_github_issue(fingerprint, sessions, session_ids)
-
-            # Mark sessions with loop_detected=True
-            for s in sessions:
-                try:
-                    ec = s.extra_context or {}
-                    ec["loop_detected"] = True
-                    s.extra_context = ec
-                    s.save()
-                except Exception as mark_err:
-                    logger.debug(
-                        "[failure-loop-detector] Could not mark session %s: %s",
-                        getattr(s, "session_id", "?"),
-                        mark_err,
-                    )
-
-        logger.info(
-            "[failure-loop-detector] Scanned %d failed sessions, found %d clusters (>= 3 failures)",
-            len(failed_sessions),
-            sum(1 for sessions in clusters.values() if len(sessions) >= 3),
-        )
-    except Exception:
-        logger.exception("[failure-loop-detector] Unhandled exception — skipping tick")
-
-
-def _latest_failure_reason(session) -> str:
-    """Extract the latest failure reason from session_events lifecycle entries.
-
-    Production code writes failure reasons via ``log_lifecycle_transition()``
-    as ``{"event_type": "lifecycle", "text": "{old}→{new}: {reason}"}`` entries
-    in ``session.session_events``.  This helper reads that path as a fallback
-    when ``extra_context`` and ``failed_reason`` are both empty.
-
-    Parse contract:
-    - Split on ``": "`` (two-char separator) with ``maxsplit=1`` to handle
-      embedded colons in the reason text (e.g. ``"health check: no progress"``).
-    - Return ``parts[1]`` when the split yields exactly 2 parts; return ``""``
-      otherwise (no separator → no reason recorded).
-
-    Guard contract:
-    - Uses ``isinstance(events, list)`` — NOT ``is None`` — because a bare
-      ``MagicMock`` auto-creates a truthy attribute that would pass a None check
-      but is not iterable as expected.
-
-    Empty-reason contract:
-    - Returns ``""`` for all malformed inputs so callers can treat a falsy return
-      as "no reason available" without special-casing.
-    """
-    events = getattr(session, "session_events", None)
-    if not isinstance(events, list):
-        return ""
-    for item in reversed(events):
-        if not isinstance(item, dict):
-            continue
-        if item.get("event_type") == "lifecycle":
-            text = item.get("text", "")
-            parts = text.split(": ", 1)
-            return parts[1] if len(parts) == 2 else ""
-    return ""
-
-
-def _compute_fingerprint(session) -> str:
-    """Compute a short error fingerprint for a session.
-
-    Resolution order for the error message component:
-    1. ``extra_context["error_message"]`` / ``extra_context["failed_reason"]``
-    2. Top-level ``session.failed_reason``
-    3. Latest lifecycle reason from ``session.session_events`` (``"{old}→{new}: {reason}"``)
-    4. Falls back to ``""`` → raw becomes ``"unknown:"`` → degenerate hash
-       ``06f5940a02173ba1`` (SHA256("unknown:")[:16]).
-
-    The session_events fallback (step 3) means that sessions whose failure
-    reason is stored only as a lifecycle transition event — the standard
-    production path via ``finalize_session()`` → ``log_lifecycle_transition()``
-    — now produce distinct fingerprints per distinct reason rather than all
-    collapsing into the same degenerate hash.  The 7-day per-fingerprint Redis
-    dedup key ensures one GitHub issue per novel cluster.
-
-    Uses HTTP status code or exception type for the non-message component when
-    available; otherwise ``"unknown"``.
-    """
-    ec = getattr(session, "extra_context", None) or {}
-    http_status = ec.get("http_status") or ec.get("status_code")
-    exc_type = ec.get("exception_type") or ec.get("error_type")
-    error_message = ec.get("error_message") or ec.get("failed_reason") or ""
-
-    # Also try top-level failed_reason on the model
-    if not error_message:
-        error_message = getattr(session, "failed_reason", "") or ""
-
-    # Fallback: read the latest lifecycle reason from session_events
-    if not error_message:
-        error_message = _latest_failure_reason(session)
-
-    if http_status is not None:
-        http_component = str(http_status)
-    elif exc_type:
-        http_component = str(exc_type)
-    else:
-        http_component = "unknown"
-
-    raw = f"{http_component}:{error_message[:80]}"
-    return hashlib.sha256(raw.encode()).hexdigest()[:16]
-
-
-def _file_github_issue(fingerprint: str, sessions: list, session_ids: list) -> None:
-    """File a GitHub issue for a failure loop cluster."""
-    try:
-        count = len(sessions)
-        sample_session = sessions[0]
-        ec = getattr(sample_session, "extra_context", None) or {}
-        error_message = (
-            ec.get("error_message")
-            or ec.get("failed_reason")
-            or getattr(sample_session, "failed_reason", "")
-            or _latest_failure_reason(sample_session)
-            or "Unknown error"
-        )[:200]
-
-        title = f"[failure-loop] {count}x same error fingerprint: {fingerprint}"
-        body = (
-            f"## Failure Loop Detected\n\n"
-            f"**Fingerprint:** `{fingerprint}`\n"
-            f"**Count:** {count} failures in last 4 hours\n"
-            f"**Sample error:** {error_message}\n\n"
-            f"**Affected session IDs (up to 5):**\n"
-            + "\n".join(f"- `{sid}`" for sid in session_ids)
-            + "\n\nAuto-filed by `failure-loop-detector` reflection."
-        )
-
-        result = subprocess.run(
-            ["gh", "issue", "create", "--title", title, "--body", body, "--label", "bug"],
-            capture_output=True,
-            text=True,
-            timeout=30,
-        )
-        if result.returncode == 0:
-            logger.warning(
-                "[failure-loop-detector] Filed GitHub issue for fingerprint %s: %s",
-                fingerprint,
-                result.stdout.strip(),
-            )
-        else:
-            logger.error(
-                "[failure-loop-detector] gh issue create failed (rc=%d): %s",
-                result.returncode,
-                result.stderr.strip(),
-            )
-    except Exception as e:
-        logger.error("[failure-loop-detector] Failed to file GitHub issue: %s", e)
-
-
-def sustainability_digest() -> None:
-    """Daily health check with exception-only Telegram delivery.
-
-    Checks all health signals locally. If everything is nominal, sends
-    NOTHING — the daily all-clear is intentional noise; live status is always
-    on the dashboard. Only spins up a full agent session to report when there
-    are anomalies worth Tom's attention.
-
-    This reflection is registered with interval 86400s (daily) as a
-    function-type callable, so the silence-on-healthy contract is deterministic
-    rather than relying on an LLM agent to choose to send nothing.
-    """
-    try:
-        from models.agent_session import AgentSession
-
-        r = _get_redis()
-        project_key = _get_project_key()
-        throttle_key = f"{project_key}:sustainability:throttle_level"
-        pause_key = f"{project_key}:sustainability:queue_paused"
-        fingerprints_key = f"{project_key}:sustainability:seen_fingerprints"
-
-        # Gather state
-        throttle_level = (r.get(throttle_key) or b"none").decode()
-        queue_paused = bool(r.exists(pause_key))
-        failure_cluster_count = r.scard(fingerprints_key) or 0
-
-        # Check circuits
-        circuits_ok = True
-        try:
-            from bridge.health import get_health
-            from bridge.resilience import CircuitState
-
-            health = get_health()
-            for name in ("anthropic", "telegram", "redis"):
-                cb = health.get(name)
-                if cb is not None and cb.state != CircuitState.CLOSED:
-                    circuits_ok = False
-                    break
-        except Exception:
-            circuits_ok = False
-
-        # Count failed sessions in last 24h
-        failed_24h = 0
-        try:
-            all_sessions = list(AgentSession.query.filter(project_key=project_key))
-            cutoff = time.time() - 86400
-            for s in all_sessions:
-                if s.status not in ("failed", "abandoned"):
-                    continue
-                completed_at = getattr(s, "completed_at", None)
-                if completed_at is None:
-                    continue
-                ts = (
-                    completed_at
-                    if isinstance(completed_at, int | float)
-                    else getattr(completed_at, "timestamp", lambda: 0)()
-                )
-                if ts >= cutoff:
-                    failed_24h += 1
-        except Exception:
-            failed_24h = -1  # unknown — treat as anomaly
-
-        # Determine if everything is nominal
-        is_nominal = (
-            circuits_ok
-            and throttle_level == "none"
-            and not queue_paused
-            and failure_cluster_count == 0
-            and failed_24h == 0
-        )
-
-        if is_nominal:
-            # Exception-only delivery: on a healthy day, send NOTHING. The daily
-            # all-clear is intentional noise Tom does not want — live status is
-            # always on the dashboard (localhost:8500). Sending nothing is success.
-            logger.info("[system-health-digest] All nominal — staying silent (no all-clear ping)")
-        else:
-            # Something noteworthy — spin up an agent to investigate and report
-            anomalies = []
-            if not circuits_ok:
-                anomalies.append("one or more service circuits are not healthy")
-            if throttle_level != "none":
-                anomalies.append(f"throttle level is {throttle_level}")
-            if queue_paused:
-                anomalies.append("queue is paused")
-            if failure_cluster_count > 0:
-                anomalies.append(f"{failure_cluster_count} active failure cluster(s)")
-            if failed_24h > 0:
-                anomalies.append(f"{failed_24h} failed/abandoned session(s) in last 24h")
-            elif failed_24h < 0:
-                anomalies.append("could not count failed sessions")
-
-            command = (
-                "You are generating the daily sustainability digest for the Valor AI system. "
-                "There are anomalies that need reporting:\n"
-                + "\n".join(f"- {a}" for a in anomalies)
-                + "\n\n"
-                "Investigate each anomaly. Collect details:\n"
-                "1. Circuit state per dependency (anthropic, telegram, redis)\n"
-                "2. Current throttle level and queue paused status from Redis\n"
-                "3. Session counts and failure details from last 24 hours\n"
-                "4. Active failure cluster count\n\n"
-                "When reporting circuit states, translate as follows"
-                " — never output the raw state string:\n"
-                "- 'closed' or 'CLOSED' → OK\n"
-                "- 'open' or 'OPEN' → DOWN\n"
-                "- 'half_open' or 'HALF_OPEN' → RECOVERING\n\n"
-                "Format as a concise Telegram message highlighting what's wrong. "
-                "Send via valor-telegram to the 'Eng: Valor' chat. "
-                "Subject line: ⚠️ Daily Health Digest — anomalies detected."
-            )
-
-            AgentSession.create_and_enqueue(
-                project_key=project_key,
-                message_text=command,
-                session_type="eng",
-                priority="low",
-                extra_context={"digest_type": "sustainability_digest"},
-            )
-            logger.info(
-                "[system-health-digest] Anomalies detected (%s) — enqueued agent session",
-                ", ".join(anomalies),
-            )
-    except Exception:
-        logger.exception("[system-health-digest] Unhandled exception — skipping tick")
